Moved `benchmark/` folder into `tools/`
